requirements.txt
amqp==2.5.2
asgiref==3.2.3
backcall==0.1.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
Django==3.0.3
django-cors-headers==3.2.1
django-sendgrid-v5==0.8.1
djangorestframework==3.11.0
future==0.18.2
idna==2.8
importlib-metadata==1.5.0
ipdb==0.12.3
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
kombu==4.6.7
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
python-dotenv==0.10.5
python-http-client==3.2.4
pytz==2019.3
redis==3.4.1
requests==2.22.0
sendgrid==6.1.1
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
zipp==2.1.0
